converted to doubles

+++ config.h
@@ -13,7 +13,7 @@ typedef uint8_t BYTE;
 typedef uint8_t bool;
 
 typedef struct {
-	float width, height;
+	double width, height;
 } Size;
 
 typedef struct {
@@ -21,14 +21,14 @@ typedef struct {
 } Color;
 
 typedef struct {
-	float x, y;
+	double x, y;
 } Vector;
 
 typedef struct Position {
-	float x;
-	float y;
-	float direction;  // angle
-	float dDir;  // change in angle from previous frame
+	double x;
+	double y;
+	double direction;  // angle
+	double dDir;  // change in angle from previous frame
 	Vector d;  // current directional vector
 } Position;
